+++ docs/Changelog.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## X.X.X (TBA)
-- TBA
+- The `networking` property of `WMTOperations`, `WMTPush`, `WMTInbox`, and `WMTOIDC` is now `public`.
+- Added `requestInterceptors` parameter to `WultraMobileToken.init(...)` and `PowerAuthSDK.createWultraMobileToken(...)`, `nil` by default.
 
 ## 3.0.0
 - Bumped dependency `PowerAuth2` to `2.0.x` (PowerAuth Mobile SDK 2.0). The `PowerAuthCore` module is no longer required and is no longer linked.